requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.0.3
oracledb==2.0.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic[email]==2.10.1